backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
pypdf
openai
python-dotenv
beautifulsoup4
sentence-transformers
numpy
